# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Infrastructure Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: serverId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Create a new deployment
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /organizations/{orgId}/deployments/{deploymentId}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get deployment details
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /organizations/{orgId}/deployments/{deploymentId}/approve:
    post:
      operationId: approveDeployment
      tags: [deployments]
      summary: Approve or reject a deployment plan
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentApproval"
      responses:
        "200":
          description: Approval result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /organizations/{orgId}/deployments/{deploymentId}/execute:
    post:
      operationId: executeDeployment
      tags: [deployments]
      summary: Execute an approved deployment
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /organizations/{orgId}/deployments/{deploymentId}/rollback:
    post:
      operationId: rollbackDeployment
      tags: [deployments]
      summary: Execute rollback for a failed deployment
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rollback started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /organizations/{orgId}/deployments/{deploymentId}/logs:
    get:
      operationId: getDeploymentLogs
      tags: [deployments]
      summary: Get deployment execution logs
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeploymentLog"
  /organizations/{orgId}/deployments/{deploymentId}/steps:
    get:
      operationId: getDeploymentSteps
      tags: [deployments]
      summary: Get deployment steps with status
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
        - name: deploymentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeploymentStep"
  /organizations/{orgId}/deployments/recent:
    get:
      operationId: getRecentDeployments
      tags: [deployments]
      summary: Get recent deployments activity feed
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Recent deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
  /organizations/{orgId}/deployments/stats:
    get:
      operationId: getDeploymentStats
      tags: [deployments]
      summary: Get deployment statistics summary
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentStats"

  # ─── AI ASSISTANT ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    DeploymentInput:
      type: object
      required: [serverId, templateId, name]
      properties:
        serverId:
          type: integer
        templateId:
          type: integer
        name:
          type: string
        description:
          type: string
        configOverrides:
          type: object
          additionalProperties: true
        scheduledAt:
          type: ["string", "null"]
          format: date-time
    DeploymentApproval:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        comment:
          type: string
    Deployment:
      type: object
      required: [id, name, status, serverId, templateId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, planning, planned, awaiting_approval, approved, rejected, running, completed, failed, rolling_back, rolled_back]
        serverId:
          type: integer
        serverName:
          type: string
        serverHost:
          type: string
        templateId:
          type: integer
        templateName:
          type: string
        templateSoftware:
          type: string
        createdById:
          type: integer
        createdByName:
          type: string
        approvedById:
          type: ["integer", "null"]
        approvedByName:
          type: ["string", "null"]
        approvalComment:
          type: ["string", "null"]
        plan:
          $ref: "#/components/schemas/DeploymentPlan"
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        duration:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    DeploymentPlan:
      type: object
      required: [summary, steps, risks, estimatedDuration, rollbackStrategy]
      properties:
        summary:
          type: string
        whyRequired:
          type: string
        benefits:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            $ref: "#/components/schemas/RiskItem"
        dependencies:
          type: array
          items:
            type: string
        estimatedDuration:
          type: integer
        estimatedDowntime:
          type: integer
        requiredPorts:
          type: array
          items:
            type: integer
        requiredPackages:
          type: array
          items:
            type: string
        backupRecommendations:
          type: array
          items:
            type: string
        rollbackStrategy:
          type: string
        expectedOutcome:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/PlanStep"
    RiskItem:
      type: object
      required: [description, severity]
      properties:
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        mitigation:
          type: string
    PlanStep:
      type: object
      required: [order, name, description, command, estimatedSeconds]
      properties:
        order:
          type: integer
        name:
          type: string
        description:
          type: string
        command:
          type: string
        whyRequired:
          type: string
        estimatedSeconds:
          type: integer
        rollbackCommand:
          type: string
    DeploymentLog:
      type: object
      required: [id, deploymentId, timestamp, level, message]
      properties:
        id:
          type: integer
        deploymentId:
          type: integer
        timestamp:
          type: string
          format: date-time
        level:
          type: string
          enum: [info, warn, error, success, debug]
        message:
          type: string
        stepName:
          type: ["string", "null"]
        output:
          type: ["string", "null"]
    DeploymentStep:
      type: object
      required: [id, deploymentId, order, name, status]
      properties:
        id:
          type: integer
        deploymentId:
          type: integer
        order:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        command:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, running, completed, failed, skipped]
        output:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        duration:
          type: ["integer", "null"]
    DeploymentStats:
      type: object
      required: [total, completed, failed, running, successRate]
      properties:
        total:
          type: integer
        completed:
          type: integer
        failed:
          type: integer
        running:
          type: integer
        rolledBack:
          type: integer
        successRate:
          type: number
        avgDuration:
          type: number
        byStatus:
          type: object
          additionalProperties: true
        byDay:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              count:
                type: integer
              failed:
                type: integer

    # ─── AI ────────────────────────────────────────────────────────────────────
...